Move agent/subscribe -> agent/rpc/subscribe
