+++ agent/rpc/subscribe/auth.go
