+++ agent/rpc/subscribe/subscribe_test.go
